make TIOCGDEV 64bit clean
suse-commit: 01973522b07f150377be51dfd636c30994ecb270

+++ drivers/char/tty_io.c
@@ -1892,8 +1892,8 @@ int tty_ioctl(struct inode * inode, struct file * file,
 		 */
 		case TIOCGDEV:
 		{
-			unsigned int ret = old_encode_dev(tty_devnum(real_tty));
-			return put_user(ret, (unsigned int*) arg);
+			unsigned long ret = old_encode_dev(tty_devnum(real_tty));
+			return put_user(ret, (unsigned long*) arg);
 		}
 
 	}
